refactor(midnight-sdk): split ratifier utilities
